New makefile! DO NOT OVERWRITE this makefile

+++ Makefile
@@ -1,47 +1,76 @@
 CC = g++
 LD = $(CC)
-
-TARGET = game
+GITBINARY = git
+FIRMWAREREPO = simulator_libraries
+FEHURL = code.osu.edu
 
 CPPFLAGS = -MMD -MP -Os -DOBJC_OLD_DISPATCH_PROTOTYPES -g
 
-IGNORED_WARNINGS = -Wall
+WARNINGS = -Wall
 
-LIB_DIR = Libraries
+LIB_DIR = simulator_libraries
 
 INC_DIRS = -I$(LIB_DIR) -I.
 
-OBJS = $(LIB_DIR)/FEHLCD.o $(LIB_DIR)/FEHRandom.o $(LIB_DIR)/FEHSD.o $(LIB_DIR)/tigr.o $(LIB_DIR)/FEHImages.o 
+OBJS = $(LIB_DIR)/FEHLCD.o $(LIB_DIR)/FEHRandom.o $(LIB_DIR)/FEHSD.o $(LIB_DIR)/FEHUtility.o $(LIB_DIR)/tigr.o $(LIB_DIR)/FEHImages.o 
 
 ifeq ($(OS),Windows_NT)
 	LDFLAGS = -lopengl32 -lgdi32
 	EXEC = game.exe
+	SHELL := CMD
 else
 	LDFLAGS = -framework OpenGL -framework Cocoa
 	EXEC = game.out
 endif
 
-main: main.o $(OBJS)
-	$(CC) $(CPPFLAGS) $(INC_DIRS) $(OBJS) main.o -o $(EXEC) $(LDFLAGS) $(IGNORED_WARNINGS)
+SRC_FILES := $(wildcard ./*.cpp)
+OBJ_FILES := $(patsubst ./%.cpp,./%.o,$(SRC_FILES))
 
-main.o: main.cpp
-	$(CC) $(IGNORED_WARNINGS) $(INC_DIRS) -c main.cpp
+all: pre-build $(EXEC)
 
-FEHLCD.o: $(LIB_DIR)/FEHLCD.cpp $(LIB_DIR)/FEHLCD.h $(LIB_DIR)/FEHUtility.o
-	$(CC) $(IGNORED_WARNINGS) $(INC_DIRS) -c $(LIB_DIR)/FEHLCD.cpp
+pre-build:
+ifeq ($(OS),Windows_NT)	
+# check for internet connection
+	@ping -n 1 -w 1000 $(FEHURL) > NUL & \
+	if errorlevel 1 \
+	( \
+		( echo "Warning: No internet connection!" ) \
+	) \
+	else \
+	( \
+		( \
+			cd $(LIB_DIR) && \
+			$(GITBINARY) stash && \
+			$(GITBINARY) pull && \
+			cd .. \
+		) \
+	) 
+else
+# Mac/Linux
+	@ping -c 1 -W 1000 $(FEHURL) > /dev/null ; \
+	if [ "$$?" -ne 0 ]; then \
+		echo "Warning: No internet connection!"; \
+	else \
+		cd $(LIB_DIR) ; \
+		$(GITBINARY) stash ; \
+		$(GITBINARY) pull ; \
+		cd .. ; \
+	fi \
 
-FEHUtility.o: $(LIB_DIR)/FEHUtility.cpp $(LIB_DIR)/FEHUtility.h
-	$(CC) $(IGNORED_WARNINGS) $(INC_DIRS) -c $(LIB_DIR)/FEHUtility.cpp
+endif
 
-FEHRandom.o: $(LIB_DIR)/FEHRandom.cpp $(LIB_DIR)/FEHRandom.h
-	$(CC) $(IGNORED_WARNINGS) $(INC_DIRS) -c $(LIB_DIR)/FEHRandom.cpp
+$(EXEC): $(OBJ_FILES) $(OBJS)
+	$(CC) $(CPPFLAGS) $(WARNINGS) $(INC_DIRS) $(OBJ_FILES) $(OBJS) -o $(EXEC) $(LDFLAGS)
 
-FEHSD.o: $(LIB_DIR)/FEHSD.cpp $(LIB_DIR)/FEHSD.h
-	$(CC) $(IGNORED_WARNINGS) $(INC_DIRS) -c $(LIB_DIR)/FEHSD.cpp
-
-tigr.o: $(LIB_DIR)/tigr.c $(LIB_DIR)/tigr.h
-	$(CC) $(IGNORED_WARNINGS) $(INC_DIRS) -c $(LIB_DIR)/tigr.c
+./%.o: ./%.cpp
+	$(CC) $(CPPFLAGS) $(WARNINGS) $(INC_DIRS) -c -o $@ $<
 
 clean:
+ifeq ($(OS),Windows_NT)
+	del $(LIB_DIR)\*.o
+	del $(LIB_DIR)\*.d
+	del *.o *.d $(EXEC)
+else
 	rm $(LIB_DIR)/*.o $(LIB_DIR)/*.d
-	rm *.o $(EXEC)
+	rm *.o *.d $(EXEC)
+endif
